Use a breadth first pathfinding algorithmn
Dump the old poorly implemented A* in favour of a simple step count,
built in a breadth first manner. This builds from destination to source,
and can be much more easily extended in the future to handle pathfinding through
teleporters and other such neat things.
also some mild-refactoring, a few off-by-one fixes for mergewaypoints,
and a heatmap debug output, currently requires AI and VERBOSE output debug channels.

haven't tested for performance, but it's likely better, as it makes the
smoothing step we had before redundant